requirements.txt
hvac==0.11.0
PyYAML==5.4.1
requests==2.26.0
click==8.0.1
